Correction of return type of math function

+++ util/maths.h
@@ -391,7 +391,7 @@ static inline float maths_interpolate(float x, float x1, float x2, float y1, flo
 	}
 }
 
-static inline int8_t maths_sign(float x)
+static inline float maths_sign(float x)
 {
 	if (x >= 0.0f)
 	{
